Revert "Switch endpoints overviewPre->overview (#2027)"
This reverts commit 9df3bd63ebdcc4a1dcbf23d23e34a021da79096f.
