+++ defi/src/adaptors/handlers/notifyStatus.ts
@@ -1,4 +1,4 @@
-import { handler } from "./processProtocolsSummary";
+import { handler } from "./getOverview";
 import { APIGatewayProxyEvent } from "aws-lambda";
 import { sendDiscordAlert } from "../utils/notify";
 import { autoBackfill } from "../cli/backfillUtilities/backfillFunction";
